pyproject.toml
[project]
name = "gen-ai-exchange"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.2.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "ipynb>=0.5.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.28",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.32",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "langgraph-cli[inmem]>=0.4.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "tldextract>=5.3.0",
    "uvicorn>=0.35.0",
    "wikipedia>=1.4.0",
]
